Move GuiSettings.java to commons
