+++ src/main/java/spleetwaise/address/ui/MainWindow.java
@@ -10,8 +10,8 @@ import javafx.scene.input.KeyCombination;
 import javafx.scene.input.KeyEvent;
 import javafx.scene.layout.StackPane;
 import javafx.stage.Stage;
-import spleetwaise.address.commons.core.GuiSettings;
 import spleetwaise.address.logic.Logic;
+import spleetwaise.commons.core.GuiSettings;
 import spleetwaise.commons.core.LogsCenter;
 import spleetwaise.commons.exceptions.SpleetWaiseCommandException;
 import spleetwaise.commons.logic.commands.CommandResult;
