+++ src/test/java/spleetwaise/commons/core/GuiSettingsTest.java
@@ -1,4 +1,4 @@
-package spleetwaise.address.commons.core;
+package spleetwaise.commons.core;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
