+++ src/test/java/spleetwaise/commons/storage/JsonUserPrefsStorageTest.java
@@ -11,8 +11,8 @@ import java.util.Optional;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.io.TempDir;
 
-import spleetwaise.address.commons.core.GuiSettings;
 import spleetwaise.address.testutil.Assert;
+import spleetwaise.commons.core.GuiSettings;
 import spleetwaise.commons.exceptions.DataLoadingException;
 import spleetwaise.commons.model.UserPrefs;
