+++ src/test/java/spleetwaise/commons/storage/StorageManagerTest.java
@@ -9,7 +9,6 @@ import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.io.TempDir;
 
-import spleetwaise.address.commons.core.GuiSettings;
 import spleetwaise.address.model.AddressBook;
 import spleetwaise.address.model.AddressBookModel;
 import spleetwaise.address.model.AddressBookModelManager;
@@ -17,6 +16,7 @@ import spleetwaise.address.model.ReadOnlyAddressBook;
 import spleetwaise.address.model.person.Person;
 import spleetwaise.address.storage.JsonAddressBookStorage;
 import spleetwaise.address.testutil.TypicalPersons;
+import spleetwaise.commons.core.GuiSettings;
 import spleetwaise.commons.model.UserPrefs;
 import spleetwaise.transaction.model.ReadOnlyTransactionBook;
 import spleetwaise.transaction.model.TransactionBook;
