+++ src/test/java/systemtests/TestApp.java
@@ -10,7 +10,6 @@ import java.util.function.Supplier;
 import javafx.stage.Screen;
 import javafx.stage.Stage;
 import spleetwaise.address.MainApp;
-import spleetwaise.address.commons.core.GuiSettings;
 import spleetwaise.address.model.AddressBook;
 import spleetwaise.address.model.AddressBookModel;
 import spleetwaise.address.model.AddressBookModelManager;
@@ -18,6 +17,7 @@ import spleetwaise.address.model.ReadOnlyAddressBook;
 import spleetwaise.address.model.person.Person;
 import spleetwaise.address.storage.JsonAddressBookStorage;
 import spleetwaise.commons.core.Config;
+import spleetwaise.commons.core.GuiSettings;
 import spleetwaise.commons.exceptions.DataLoadingException;
 import spleetwaise.commons.model.UserPrefs;
 import spleetwaise.commons.storage.UserPrefsStorage;
